refactor: consolidação de regras de negócio, auditoria e controllers
- Controllers: Implementação do TurmaController e AuditoriaController com rotas RESTful.

- Services: Refatoração de Aluno, Turma, Usuario e Falta para suportar exclusão lógica e integrações atômicas com a Auditoria.

- Repositories: Ajustes nas queries do FaltaRepository e TurmaRepository.

- Models: Atualização das classes de domínio (Turma e AcaoAuditoria) para refletir os novos comportamentos.

+++ src/main/java/com/pelizzaris/sufs/repository/FaltaAlunoRepository.java
@@ -15,9 +15,9 @@ public interface FaltaAlunoRepository extends JpaRepository<FaltaAluno, Long> {
 
     long countByAlunoIdAndFaltaDataFaltaBetween(UUID alunoId, LocalDate dataInicio, LocalDate dataFim);
 
-    List<FaltaAluno> findByAlunoTurmaId(Integer turmaId);
+    List<FaltaAluno> findByAlunoTurmaId(Long turmaId);
 
-    List<FaltaAluno> findByAlunoTurmaIdAndFaltaDataFaltaBetween(Integer turmaId, LocalDate inicio, LocalDate fim);
+    List<FaltaAluno> findByAlunoTurmaIdAndFaltaDataFaltaBetween(Long turmaId, LocalDate inicio, LocalDate fim);
 
     void deleteByFaltaId(Long id);
 }
